Put my authentication code into its own utils component
